% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIShouse_by_Floor.R
\docType{data}
\name{DIShouse_by_Floor}
\alias{DIShouse_by_Floor}
\title{Number of Housing Units by Main Material Used for Floor Construction According to District}
\format{
A data frame.
}
\source{
Department of Census and Statistics, Sri Lanka.
Censuses of Population and Housing 2012 and 2024.
}
\usage{
DIShouse_by_Floor
}
\description{
Number of housing units classified by the main material used for floor
construction at the district level in Sri Lanka based on the 2012 and 2024
Censuses of Population and Housing.
}
\examples{
head(DIShouse_by_Floor)
}
\keyword{datasets}
