% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DISpop_by_FiveYearAgeGroup.R
\docType{data}
\name{DISpop_by_FiveYearAgeGroup}
\alias{DISpop_by_FiveYearAgeGroup}
\title{Population by Five-Year Age Groups According to District}
\format{
A data frame.
}
\source{
Department of Census and Statistics, Sri Lanka.
Censuses of Population and Housing 2012 and 2024.
}
\usage{
DISpop_by_FiveYearAgeGroup
}
\description{
Distribution of the population by five-year age groups at the district level
in Sri Lanka based on the 2012 and 2024 Censuses of Population and Housing.
}
\examples{
head(DISpop_by_FiveYearAgeGroup)
}
\keyword{datasets}
