% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_pop_age_sex.R
\docType{data}
\name{gn_pop_age_sex}
\alias{gn_pop_age_sex}
\title{Population by Sex and Age According to Districts}
\format{
A data frame.
}
\source{
Department of Census and Statistics Sri Lanka
}
\usage{
gn_pop_age_sex
}
\description{
Census of Population and Housing 2024.
}
\keyword{datasets}
