% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNpop_by_EthnicGroup.R
\docType{data}
\name{GNpop_by_EthnicGroup}
\alias{GNpop_by_EthnicGroup}
\title{Population by Ethnic Group According to Grama Niladhari Division}
\format{
A data frame.
}
\source{
Department of Census and Statistics, Sri Lanka.
Census of Population and Housing 2024.
}
\usage{
GNpop_by_EthnicGroup
}
\description{
Distribution of population by ethnic group across Grama Niladhari Divisions (GND) in Sri Lanka
based on the 2024 Census of Population and Housing.
}
\keyword{datasets}
