% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNpop_by_Religion.R
\docType{data}
\name{GNpop_by_Religion}
\alias{GNpop_by_Religion}
\title{Population by Religion According to Grama Niladhari Division}
\format{
A data frame.
}
\source{
Department of Census and Statistics, Sri Lanka.
Census of Population and Housing 2024.
}
\usage{
GNpop_by_Religion
}
\description{
Distribution of population by religion across Grama Niladhari Divisions (GND) in Sri Lanka
based on the 2024 Census of Population and Housing.
}
\examples{
head(GNpop_by_Religion)
}
\keyword{datasets}
